requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.10
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-import-export==1.2.0
django-reversion==3.0.2
djangorestframework==3.9.1
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
mysqlclient==1.4.1
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.4.1
protobuf==3.6.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==4.2b1
requests==2.21.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
